Update issue 11
I've worked at the bug we mentioned this morning and then found that when user input a new Batch record, the format of bttime stored in database is not correct. To be specific I added a new batch just right now, its btime is saved as 2011-08-18T00:00:00, there is a redundant 'T' in the data, which is why the statistic function dose not run successfully. For the Batch function is not owned by me, I am afraid if I modify your code will cause another problems, so pls hv a check about this issue.