Update issue 51
TIt is very helpful to use JFreeChart for reporting purposes of your system/ programs. However, you may notice that once you incorporate this in your system and you close it, your whole system/ application/ program will also exit or be closed. This is because JFreeCharts usually extends APPLICATIONFRAME that LISTENS to the close button and CLOSES EVERYTHING when clicked. (which is horrible for your system).

Solution?
Just override the windowclosing function!
Like this:

public void windowClosing(final WindowEvent evt){
if(evt.getWindow() == this){
dispose();

}
}